[semantic-arc] When deserializing functions, only set the unqualified bit when we deserialize the body.
There is no reason to make a declaration have qualified or unqualified
ownership. Incidently, this also fixes a bug where we were only setting the
unqualified bit for functions created in the deserializer. If we created a
declaration and linked it in later, we would not get the proper setting for
ownership qualification.

rdar://28851920