[build-script-impl]: Fixes version number on Linux from 3.0 to 3.0.x
Brief:
COMPILER_VENDOR is None on Linux, so we get version 3.0 instead of 3.0.x.

Full:
On Linux, SWIFT_VERSION is unset because the ``None)`` case for ``COMPILER_VENDOR`` is empty.
When ``lib/Basic/CMakeLists.txt`` tries to match the empty ``SWIFT_VERSION`` against regexps
testing for a version triple, matching fails and we set ``DSWIFT_HAS_VERSION_PATCHLEVEL=0``.
Then, in ``lib/Basic/Version.cpp``, since PATCHLEVEL is 0, it chooses the ``#else``
and ``SWIFT_VERSION_STRING`` is set to a pair instead of a triple, which gives ``3.0``.