[DI] When checking class inits for uses of self, ignore strong_retain.
This specific issue came up due to my SILGen changes for copy_value,
destroy_value. Specifically, we used to emit the diagnostic:

  super.init isn't called on all paths before returning from initializer.

After my change, we began to emit:

  'self' used before super.init call

Since the code was:

  init() {}

there is clearly no use, so the new error is incorrect or at minimum
misleading. When I investigated why the change in diagnostic happened, it was
because as a result of my SILGen change, the order of the strong_retain and
return in the use-list swapped in the following code:

  %1 = load %0 : $*MultipleInitDerived
  strong_retain %1 : $MultipleInitDerived
  return %1 : $MultipleInitDerived

Before my change, return was visited first, causing the correct diagnostic to be
emitted. After my change, the strong_retain is visited first causing the second
diagnostic to be emitted. Since we only emit one source error for each
SILLocation, in the 2nd case, the error associated with the return is not
emitted.

We should not be considering strong_retain to be a use that propagates
liveness. The reason why is that a strong_retain should always be paired with
some local use that consumes it, whether that is an apply, closure, or a store
into memory. Since we will always have a consuming operation for the retain and
the consuming operation actually is able to be something visible at the source
level to the user (versus the retain), it makes more sense to just ignore
strong_retain.

In general though, we should try to be more careful about how we swallow errors
in DI, especially in the light of use-list twiddling. I attempted to improve our
testing situation by added a flag '-definite-init-visit-debuginfo-locs' that
would cause DI to consider SIL Location debug info locs when deciding whether or
not to swallow errors. Sadly, our diagnostics (from what I can tell) do not
support emitting diagnositics for such locations, so I failed.

Thus I could not actually test this, but it will fail once I get in the
copy_value, destroy_value change, so that /should/ act as good enough of a test.

rdar://28851920