Modernize text formatting prototype
Since we are thinking about strings I wanted to clean this up and extract any useful ideas.

I think the main idea of value here is that, when generating large amounts of text (as in a web service), or writing a large XML file, a program has to format many individual items such as integers and dates. If these are formatted into stored strings which are then written to a file, the intermediate string is immediately discarded and the work to construct and store it was wasted.

This prototype abstracts formatted representation as `TextOutputStreamable` instances that hold the underlying things being formatted, and push text directly into a `TextOutputStream` instance.

In a world with a `StringProtocol`, though, maybe we don't need `TextOutputStreamable` et al.  There's no reason that models of `StringProtocol` couldn't be built the same way.