Remove Darwin module's dependency on ctypes.h
Upstream cfe change r284797 adds explicit handling for modules
named Darwin, and requires that it not depend on modules which
it does not parent. Because ctypes is a peer module of Darwin,
this means that Darwin must no longer depend on ctypes.

This can be avoided by duplicating some integral definitions
in MacTypes.h, and by adding extra imports to a few tests.