[Parse] Change the way to add parenthesis in fix-it for deprecated protocol<..>
Fixes: https://bugs.swift.org/browse/SR-3124

"If '>' had any split token" was not really a correct condition for adding
parenthesis.

For instance:
    let a:protocol<P1, P2>=someThing
We don't need parenthesis for this case:
    let a:P1 & P2=someThing

On the other hand:
    typealias PMetaType = protocol<P1, P2>.Type
This doesn't produce a split token, but we need the parenthesis.
    typealias PMetaType = (P1 & P2).Type

The new rule is: "If the token after '>' can be any postfix TypeRepr"
(i.e. '?', '!', '.Type' or '.Protocol')