[DI] Given x' = load x; (apply (class_method x') x'), check the apply for DI violations, not class_method.
This is working around an additional use-list DI ordering issue that I exposed
when implementing High Level Memory Operations. Specifically, DI started to
error on:

  (class_method x)

instead of on:

  (apply (class_method x) x)

We would also try to emit an error on the apply, but we would squelch the apply
error (which is more accuracte) since we had already emitted the class_method
error.

This commit conservatively checks for this condition and skips the class method
so we can emit the more descriptive error on the apply.