Add additional pre-requisites needed for building on Linux
Although the build for the Swift compiler doesn't need them,
libdispatch on Linux has an autoconf and libtool dependency.

The instructions at the top-level README has a list of dependencies
which can be installed using a `sudo apt-get install` command, but
when building using just these dependencies the libdispatch project
cannot be built.

By adding the libdispatch dependencies to the top level project,
the set of dependencies can be built in one go.

Issue: SR-3222