[QoI] Improve diagnostics for type member names shadowing top-level functions
Consider expression with an implicit 'self.' reference:

extension Sequence {
  func test() -> Int {
    return max(1, 2)
  }
}

We currently shadow names that appear in nested scopes, so the
top-level two-argument min()/max() functions are shadowed by the
versions of min()/max() in Sequence (which don’t take the same
number of arguments). This patch aims to improve situation on this
front and produce better diagnostics when that happens, hinting
the user about what went wrong and where matching function is
declared.

Resolves <rdar://problem/25341015>.