[AST] Make GenericEnvironment's reverse mapping deterministic.
GenericEnvironment walked its input mapping in DenseMap order while
populating a reverse mapping from archetypes to interface
types. However, this mapping is not unique, because two generic
parameters can end up mapping to the same archetype. In such cases,
which generic parameter we mapped to was nondeterministic.

Make this deterministic by preferring to map back to the earlier
generic parameter.