[Clang Importer] Support multiple alternate decls
ImportDecl has an alternate decl quasi-escape-hatch where it can
create alternative declarations for the same original Clang node. But,
it was limited at one, which worked accidentally in the case of
VisitObjCMethodDecl's attempts at importing a factory-init-suppressed
root-class method that points to the imported-as-init Swift
declaration.

This commit changes that to a TinyPtrVector, to support the upcoming
occasional case where we have more.