[Archetype builder] Replace MapVector with parallel SmallVectors.
In the archetype builder, separate out the generic parameters (which
were the keys of the MapVector) from the potential archetypes (which
where the values of the MapVector) into two parallel
SmallVectors. This saves on storage, lets us avoid building up a
SmallVector of GenericTypeParamTypes, and helps flesh out this pattern
as a way of eliminating DenseMaps in favor of arrays.