Adapt to llvm NamedRegionTimer changes
llvm requires a Name/Description pair for NamedRegionTimers as of
r287369. Because most of the previously used "names" for timers are
actually longer descriptions containing multiple words and spaces. I
added a Name parameter to the llvm side to produce machine parseable
output with names that rather resemble a programming a language
identifier rather than a sentence.

This is the easiest possible fix that uses the existing string for
name+description of the timer just to be compatible with the new API.
It would be good if someone familiar with swift could add proper short
names to the timers.