[AST] Tail-allocate protocol and superclass storage for ArchetypeType.
Superclass constraints are uncommon, so put them into trailing-object
storage. Similarly, there's no reason to pay for the pointer + size of
the array of protocols to which the archetype conforms when we can use
trailing objects and existing padding bits.

Note that the nested types are *not* tail-allocated because they will
(eventually) be lazily allocated.