[AST] Lazily expand the nested types of an ArchetypeType.
An ArchetypeType has all of the information it needs to lazily compute
its set of nested types (hint: they're the associated types of all of
the protocols to which it conforms). Do so only when asked about the
nested types.

Not that there are two levels of laziness happening here: first, we
only create the (sorted) array of nested types when asked about the
nested types the first time. Second, even when we do create it, we
only record the names of these types---the actual archetypes (or
concrete types) behind these names are lazily created when requested.

Of course, all of this laziness is short-lived, because
ArchetypeBuilder::getGenericEnvironment() forces all of the archetypes
to be constructed.